+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added base class for protocol and message parsers.
 - Added Ethernet protocol parser.
 - Fixed VLAN Tag attribute setters.
+- Added IPv4 protocol parser.